requirements.txt
boto3==1.17.72
crispy-bootstrap5==0.3.1
dj-database-url==0.5.0
django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-oauth-toolkit==1.5.0
django-otp==1.0.5
django-recaptcha==2.0.6
django-redis==4.12.1
django-storages==1.11.1
djangorestframework==3.12.4
drf-nested-routers==0.93.3
gunicorn==20.1.0
httptools==0.2.0
pillow==8.2.0
psycopg2-binary==2.8.6
qrcode==6.1
sentry-sdk==1.1.0
social-auth-app-django==4.0.0
uvicorn==0.13.4
uvloop==0.15.2
whitenoise==5.2.0
